### 2022-07-20 903b069
Updating table for badges on READMD.rst
