+++ README.rst
@@ -5,15 +5,12 @@ Python-CPD
 .. |master_badge| image:: https://github.com/siavashk/pycpd/actions/workflows/build-test.yml/badge.svg?branch=master
 .. |development_badge| image:: https://github.com/siavashk/pycpd/actions/workflows/build-test.yml/badge.svg?branch=development
 
-+=================+=====================+
++-----------------+---------------------+
 | master          | |master_badge|      |
 +-----------------+---------------------+
 | development     | |development_badge| |
 +-----------------+---------------------+
 
-
-
-.. image:: https://github.com/siavashk/pycpd/actions/workflows/build-test.yml/badge.svg?branch=development
 
 `Documentation <https://siavashk.github.io/pycpd/>`_
 
