### 2022-07-29 #4585 @Kevin-Mattheus-Moerman
@gattia our board has decided this may be in scope for JOSS. We will therefore proceed to assign a handling editor here. However, we have a bit of a backlog at the moment so I will waitlist this for now until an editor becomes available.
