### 2022-08-17 #4681 @editorialbot
[REVIEW]: PyCPD: Pure NumPy Implementation of the Coherent Point Drift Algorithm
**Submitting author:** @gattia (<a href="http://orcid.org/0000-0001-6717-8979">Anthony Gatti</a>)
**Repository:** https://github.com/siavashk/pycpd
**Branch with paper.md** (empty if default branch): master
**Version:** v2.0.1-JOSS-Paper
