### 2022-08-17 siavashk/pycpd#71 @rickie95
[Documentation] Add contribution file with guidelines
Part of reviewing process for JOSS acceptance (openjournals/joss-reviews/issues/4681)

I couldn't find clear instructions on how to participate in this project. Ideally you should provide a CONTRIBUTE file with some informations about:

- **How to contribute to the software** (e.g. how to create a pull request or minimum requirements for accept a PR )
- **How to report issues**: your procedure of choice for handle bug reports (is there a template to follow? is required to provide a minimum reproducible example? is ok just a simple issue opened on GitHub?)
- **How to reach support**: alternate channels where to get help (email, forum, ecc..)
