### 2022-08-30 #4681 @zexinyang
Hi @hugoledoux, @gattia,

I've finished my review. In this work, the authors developed a NumPy version of the well-known CPD point cloud registration algorithm. Overall, it is a handy piece of software that has already drawn the attention of the community. The software paper is well written, but typos should be corrected and references should be added (see comments below). I recommend accepting the work after a minor revision. Please find below detailed comments.

### Software paper
- Line 29: please cite references w.r.t. "other previous registration methods".
- Line 30: "didnt" => "didn't"
- Line 35: "maximze" => "maximize"
- Line 45-46: the Matlab version of the CPD algorithm needs to be cited, and the [C++ implementation](https://github.com/gadomski/cpd) created on 09 October 2014 (earlier than the presented software) should be mentioned and cited.
- In Figure 1, the caption should clarify the meaning of "Q" in subfigures.

### Documentation
- **A statement of need** and **community guidelines** are needed but currently lacking.
- **Installation instructions** can be made more clear by reorganizing them as follows.

  > ## Installation
  > ### Install from PyPI
  > ```
  > pip install pycpd
  > ```
  >
  > ### Install from Source
  > Clone the repository to a location, referred to as the `root` folder. For example:
  > ```
  > git clone https://github.com/siavashk/pycpd.git $HOME/pycpd
  > ```
  > Install the package:
  > ```
  > pip install .
  > ```
  > or
  > ```
  > make requirements
  > make build
  > ```
  > ### Install Matplotlib for Visualization
  > For running sample registration examples under examples, you will need Matplotlib to visualize the registration. This can be downloaded by running:
  > ```
  > pip install matplotlib
  > ```
  > or
  > ```
  > make visualize
  > ```
- **Examples** can be improved by adding a brief tutorial on registering custom data and providing hints on how to tune the parameters. Besides, I believe that the current fish data are unsuitable for testing rigid or affine registration algorithms since the transformation between them is deformable (as shown in the below figure). It would be much more intuitive to provide fish data with rigid/affine transformations for the rigid/affine registration examples.
- **Functionality documentation** can be improved by including example inputs and outputs.

### Errors and warnings
[Two minor issues](https://github.com/siavashk/pycpd/issues/72#issue-1355421231), which appear to be easily resolved, have been found when verifying the functionality of the presented software.
