### 2022-11-04 #4681 @gattia
Thanks for the note. And sorry for the delay. I was mostly waiting for R1 to come in before doing everything at once, but I think I now see that they just completed the checklist and have no comments?

There isn't much in there, I'll get the response done in the next week.

Anthony
