### 2022-11-04 #4681 @hugoledoux
@rickie95 I see you clicked some of the checkboxes above, but nothing posted here as feedback. Could you please let us what you thought of the software (good and less good), and why some checkboxes were not clicked?
