### 2022-11-05 e8d614f
Updating JOSS paper.md based on feedback: https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1231407510
Now completed recommendations in "Software paper"
