+++ JOSS_Submission/paper.bib
@@ -66,3 +66,18 @@ abstract = {We study the problem of creating a complete model of a physical obje
     year={2017}, 
     month={May}
 } 
+
+@software{MATLAB,
+  author = {The MathWorks Inc.},
+  title = {MATLAB},
+  url = {https://www.mathworks.com/help/vision/ref/pcregistercpd.html},
+  version = {version 9.5 (R2018b)},
+  year = {2018},
+}
+
+
+@misc{gadomski, 
+  title={Gadomski/CPD: C++ implementation of the coherent point drift point set registration algorithm.}, 
+  url={https://github.com/gadomski/cpd}, 
+  author={Gadomski, Pete}
+} 
