### 2022-11-05 71cc540
Make installation recommendations from JSS review https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1231407510
Documentation/Installation instructions can be made more clear by reorganizing them as follows.
