### 2022-11-05 08e8cf0
Update Readme & info re: contributing & community. - Added a CONTRIBUTING.MD document to outline how to contribute. - Added content to Readme that points to the CONTRIBUTING.MD. - Added content to Readme that points to the CODE_OF_CONDUCT.md - Added ISSUE_TEMPLATES to help with submitted issues.
- Resolving issue:
https://github.com/siavashk/pycpd/issues/71

- Resolving comments about `community guidelines` in reviewer comment here:
https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1231407510
