### 2022-11-05 05288cd
Update affine/rigid examples: - Provided option to make examples truly rigid/affine. - Set default examples to run these truly rigid/affine examples.
