+++ examples/fish_rigid_3D.py
@@ -16,14 +16,26 @@ def visualize(iteration, error, X, Y, ax):
     plt.pause(0.001)
 
 
-def main():
+def main(true_rigid=True):
     fish_target = np.loadtxt('data/fish_target.txt')
-    X = np.zeros((fish_target.shape[0], fish_target.shape[1] + 1))
-    X[:, :-1] = fish_target
+    X1 = np.zeros((fish_target.shape[0], fish_target.shape[1] + 1))
+    X1[:, :-1] = fish_target
+    X2 = np.ones((fish_target.shape[0], fish_target.shape[1] + 1))
+    X2[:, :-1] = fish_target
+    X = np.vstack((X1, X2))
 
-    fish_source = np.loadtxt('data/fish_source.txt')
-    Y = np.zeros((fish_source.shape[0], fish_source.shape[1] + 1))
-    Y[:, :-1] = fish_source
+    if true_rigid is True:
+        theta = np.pi / 6.0
+        R = np.array([[np.cos(theta), -np.sin(theta), 0], [np.sin(theta), np.cos(theta), 0], [0, 0, 1]])
+        t = np.array([0.5, 1.0, 0.0])
+        Y = np.dot(X, R) + t
+    else:
+        fish_source = np.loadtxt('data/fish_source.txt')
+        Y1 = np.zeros((fish_source.shape[0], fish_source.shape[1] + 1))
+        Y1[:, :-1] = fish_source
+        Y2 = np.ones((fish_source.shape[0], fish_source.shape[1] + 1))
+        Y2[:, :-1] = fish_source
+        Y = np.vstack((Y1, Y2))
 
     fig = plt.figure()
     ax = fig.add_subplot(111, projection='3d')
@@ -35,4 +47,4 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    main(true_rigid=True)
