### 2022-11-05 0122654
Adding true_rigid=True to example
