### 2022-11-05 4068f9d
Add warnings for string literals.
This is based on the request in issue:
https://github.com/siavashk/pycpd/issues/72#issue

related to JOSS review.
