### 2022-11-05 siavashk/pycpd#72 @gattia
@zexinyang thanks so much for the review, and for looking into this!
I have mostly completed all components of the review and should be submitting my response shortly.

For the errors here, I tried to re-create the errors and couldn't. In the end, it looks like you might have an old version of pycpd installed because the lines of code that have an error don't match the current code. Specifically, there should be an additional if/else statement in the `def transform_point_cloud` function that tests if you are trying to transform a low-rank version or not.

Could you try re-pulling, installing, and then running the examples/tests?

```
# activate your environment
cd /home/geo3d/Zexin/project/pycpd
git pull origin master
pip install .
make dev                        # optionally install pytest using make dev if you haven't already.
make test                        # run the tests.
python examples/fish_deformable_3D_register_with_subset_of_points.py
```
