### 2022-11-05 1294b0b
Updating Introduction of Readme.
Updaying the introduction to explicitly say who this package is for (the python community).
We also specify more informaiton about the data that the package works with (MxN arrays).

This is in response to the comment by reviewer for JOSS:
https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1231407510
