### 2022-11-05 3066941
Adding additional references to JOSS paper: This is for reviewer 2 comment: https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1231407510
