+++ JOSS_Submission/paper.bib
@@ -36,6 +36,33 @@ abstract = {We study the problem of creating a complete model of a physical obje
   doi={10.1109/TPAMI.2010.46}
 }
 
+@article{FITZGIBBON20031145,
+  title = {Robust registration of 2D and 3D point sets},
+  journal = {Image and Vision Computing},
+  volume = {21},
+  number = {13},
+  pages = {1145-1153},
+  year = {2003},
+  note = {British Machine Vision Computing 2001},
+  issn = {0262-8856},
+  doi = {https://doi.org/10.1016/j.imavis.2003.09.004},
+  url = {https://www.sciencedirect.com/science/article/pii/S0262885603001835},
+  author = {Andrew W Fitzgibbon},
+  keywords = {Iterated Closest Point, Range image registration, Levenberg–Marquardt},
+  abstract = {This paper introduces a new method of registering point sets. The registration error is directly minimized using general-purpose non-linear optimization (the Levenberg–Marquardt algorithm). The surprising conclusion of the paper is that this technique is comparable in speed to the special-purpose Iterated Closest Point algorithm, which is most commonly used for this task. Because the routine directly minimizes an energy function, it is easy to extend it to incorporate robust estimation via a Huber kernel, yielding a basin of convergence that is many times wider than existing techniques. Finally, we introduce a data structure for the minimization based on the chamfer distance transform, which yields an algorithm that is both faster and more robust than previously described methods.}
+}
+
+@INPROCEEDINGS{924423,
+  author={Rusinkiewicz, S. and Levoy, M.},
+  booktitle={Proceedings Third International Conference on 3-D Digital Imaging and Modeling}, 
+  title={Efficient variants of the ICP algorithm}, 
+  year={2001},
+  volume={},
+  number={},
+  pages={145-152},
+  doi={10.1109/IM.2001.924423}
+}
+
 @Article{         harris2020array,
  title         = {Array programming with {NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
