### 2022-11-05 e7a6bb5
Update readme example. - Add example - Add practice advice (fit on subsample of point cloud apply to whole) - Add general advice about tuning the registration parameters.
For comment in review:
https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1231407510
