### 2022-11-05 siavashk/pycpd#71 @gattia
Thank you for the comment!

Closing because the following commit adds a `CONTRIBUTING.md` with extensive documentation on how to contribute and references it in the README.

https://github.com/siavashk/pycpd/commit/08e8cf0b86018948e670cdf1d5e10b450c4547b1
