### 2022-11-05 #4681 @gattia
@zexinyang thank you very much for spending the time to thoroughly review the paper, code, tests, and other parts of the library. Your suggestions have all been great and I think improved the paper as well as the software as a whole. I have completed changes based on your recommendations on the `joss_review` branch that can be found here https://github.com/siavashk/pycpd/tree/joss_review.

Thanks again!

Anthony.

Software paper updates:
- All of the suggested changes to the manuscript were made in these two commits:
    - https://github.com/siavashk/pycpd/commit/e8d614f68b18775be562a1133af4c3215f0df48c
    - https://github.com/siavashk/pycpd/commit/3066941bff99b557111c566fb848a72ae0ba4074

- Documentation changes:
    - Adding statement of need:
        - I believe that the original "Introduction" of the README file had most of the elements necessary for the statement of need. To fully satisfy this need, I have slightly expanded this section to specify the community that the package is intended for, and have provided more information about the data the algorithm can be used for.
        - https://github.com/siavashk/pycpd/commit/1294b0b4a6d17b63557eff2ee8134c7275539b01
    - Community Guidelines:
        - Thank you for noting this omission. We had included a `CODE_OF_CONDUCT.md`, but it was never referenced. We have now added a reference to it in the Readme. We have also added a document about how to Contribute to the repository, and have linked this to the Readme page as well.
        - This is all outlined in this commit: https://github.com/siavashk/pycpd/commit/08e8cf0b86018948e670cdf1d5e10b450c4547b1
    - Installation instruction suggestions:
        - All of the suggestions were implemented in this commit: https://github.com/siavashk/pycpd/commit/71cc5407a537fd726e20edaff243df674bf3dba4
    - Examples:
        - A basic example was added to the readme page. Also, below the example are a few basic tips, recommendations, and descriptions of parameters that can be tuned for optimizing the registration.
            - These changes are included in this commit: https://github.com/siavashk/pycpd/commit/e7a6bb5a19213b1a51a02227232e869f2a1f9d29
        - As suggested, the affine and rigid registration examples were updated to actually be rigid/affine:
            - https://github.com/siavashk/pycpd/commit/05288cd0b84b9e864b453df1a7b61e703c8de6c5
    - Functionality documentation:
        - I believe that this is addressed in a few ways. First, the examples and tests all load example data from the /data folder. This is one example of how data can be formatted. Also, the newly added Example on the Readme page creates fake data and describes the format that it should take (https://github.com/siavashk/pycpd/commit/e7a6bb5a19213b1a51a02227232e869f2a1f9d29). Finally, the new updates to the introduction better describe the shapes of the input numpy arrays (https://github.com/siavashk/pycpd/commit/1294b0b4a6d17b63557eff2ee8134c7275539b01).

Issue filed one the repository:
- I made the suggested changes that will remove the warnings here:
    - https://github.com/siavashk/pycpd/commit/4068f9daebcf246412ef8da9756414442c209999
- I looked into the reported errors in the examples and the tests but couldn't reproduce them (also, in our github actions all tests are passing on linux, windows, and mac). Looking into this a bit further, it seems like the code showed in your error is an older version of pycpd, or something along those lines. I've mentioned this in reply to your issue here: (https://github.com/siavashk/pycpd/issues/72#issuecomment-1304392097). If you dont mind pulling the latest version, re-installing, and re-running the tests I think that the issues will be resolved.
