### 2022-11-06 #4681 @rickie95
> @rickie95 I see you clicked some of the checkboxes above, but nothing posted here as feedback. Could you please let us what you thought of the software (good and less good), and why some checkboxes were not clicked?

Sorry for the delay, I expected a faster pace for the dialogue with the authors, then I had to focus on other things, my bad.

About the work: I agree with the suggestions made by @zexinyang, now that they're solved I guess we can proceed with the next stage. From an overall perspective both the software and the paper are in a great shape, for the former I can also speak from an user's view.
