### 2022-11-06 #4681 @gattia
@rickie95 thank you so much for the positive review/comments, and for the Issue pulled on the repository. The addition of guidelines for contributing was a big missing piece. I think that those guidelines will reduce the barrier to others contributing to the community.
