### 2022-11-10 eba0199
Updating paper citations
