+++ JOSS_Submission/paper.bib
@@ -95,11 +95,9 @@ abstract = {We study the problem of creating a complete model of a physical obje
 } 
 
 @software{MATLAB,
-  author = {The MathWorks Inc.},
-  title = {MATLAB},
+  title = {The MathWorks Inc.},
   url = {https://www.mathworks.com/help/vision/ref/pcregistercpd.html},
   version = {version 9.5 (R2018b)},
-  year = {2018},
 }
 
 
