### 2022-11-15 b262e9b
Some fixes while I proof-read your paper
please use a spellchecker next time
