### 2022-11-15 #4681 @hugoledoux
https://github.com/siavashk/pycpd/pull/76

+ one sentence I don't understand why it,s there (or what it means):
"Examples of the PyCPD algorithm are included (Figure 1)."
