### 2022-11-15 72f2197
Fix wording re: review suggestion.
https://github.com/openjournals/joss-reviews/issues/4681#issuecomment-1315355838
