### 2022-11-15 #4681 @gattia
@hugoledoux - thanks for the edits. I've also updated the wording of that sentence to be:

"Examples of how to use the PyCPD algorithm are included in the package, **Figure 1** displays the visualization corresponding with a 3D rigid registration example"

Let me know if you are OK with this edit or if you have others and I will merge the branch with master (where the paper being reviewed currently is).
