### 2022-11-16 #4681 @hugoledoux
@gattia yes it's fine, please accept the PR

Can you also fix the DOI, as the error just above shows?

Then we can move on to officially accepting the submission.
