### 2022-11-17 #4681 @hugoledoux
We're nearly there. At this point could you:
- [ ] Accept this PR fixing 2-3 small things: https://github.com/siavashk/pycpd/pull/78
