### 2022-11-28 #4681 @gattia
Sorry for the delay in getting these last pieces wrapped up. Those items should all be done and listed below.

Let me know if there is anything else.

- The PR is merged: https://github.com/siavashk/pycpd/pull/78
- tagged version: v2.0.1-JOSS-Paper:
    https://github.com/siavashk/pycpd/releases/tag/v2.0.1-JOSS-Paper
- Zenodo DOI for the software version:
    doi.org/10.5281/zenodo.7343379
