### 2022-11-29 305b6fd
Consolidate functions used more than once - gaussian_kernel() - low_rank_eigen()
