### 2022-11-29 5a8015a
Update constrained examples to show matching points     - Make target matching point an empty circle     - Make source matching point a green star Including these matching points helps show what the matching points are, and helps identify how they are affecting the registration process.
