+++ examples/fish_constrained_deformable_3D_lowrank.py
@@ -4,11 +4,28 @@ from mpl_toolkits.mplot3d import Axes3D
 from pycpd import ConstrainedDeformableRegistration
 import numpy as np
 
+fish_source = np.loadtxt('data/fish_source.txt')
+marker_size = 100
+N_pts_include = 61
+IDs = [1,10,20,30]
+IDs_Y = IDs + [fish_source.shape[0] + i for i in IDs]
+IDs_X = IDs + [N_pts_include + i for i in IDs]
 
 def visualize(iteration, error, X, Y, ax):
     plt.cla()
-    ax.scatter(X[:, 0],  X[:, 1], X[:, 2], color='red', label='Target')
-    ax.scatter(Y[:, 0],  Y[:, 1], Y[:, 2], color='blue', label='Source')
+    ids_X = np.arange(0, X.shape[0])
+    ids_X = np.delete(ids_X, IDs_X)
+
+    ids_Y = np.arange(0, Y.shape[0])
+    ids_Y = np.delete(ids_Y, IDs_Y)
+
+
+    ax.scatter(X[ids_X, 0],  X[ids_X, 1], X[ids_X, 2], color='red', label='Target')
+    ax.scatter(Y[ids_Y, 0],  Y[ids_Y, 1], Y[ids_Y, 2], color='blue', label='Source')
+
+    ax.scatter(X[IDs_X, 0],  X[IDs_X, 1], X[IDs_X, 2], color='red', label='Target Constrained', s=marker_size, facecolors='none')
+    ax.scatter(Y[IDs_Y, 0],  Y[IDs_Y, 1], Y[IDs_Y, 2], color='green', label='Source Constrained', s=marker_size, marker=(5, 1))
+    
     ax.text2D(0.87, 0.92, 'Iteration: {:d}'.format(
         iteration), horizontalalignment='center', verticalalignment='center', transform=ax.transAxes, fontsize='x-large')
     ax.legend(loc='upper left', fontsize='x-large')
@@ -36,8 +53,9 @@ def main():
     Y = np.vstack((Y1, Y2))
 
     # select fixed correspondences
-    src_id = np.int32([1,10,20,30, len(Y1)+1, len(Y1)+10, len(Y1)+20, len(Y1)+30])
-    tgt_id = np.int32([1,10,20,30, len(X1)+1, len(X1)+10, len(X1)+20, len(X1)+30])
+    # select fixed correspondences
+    src_id = np.int32(IDs_Y)
+    tgt_id = np.int32(IDs_X)
 
     fig = plt.figure()
     ax = fig.add_subplot(111, projection='3d')
