### 2022-12-15 #4681 @gattia
Thank you @zexinyang, @rickie95 for the great feedback that improved the manuscript and software.
And thank you @hugoledoux @arfon for helping guide the process efficiently (and the patience for my response).

My first experiences with JOSS has been awesome one.

Best,

Anthony.

Special thanks to @editorialbot ! 😄
